requirements.txt
aiohttp>=3.6.0,<4
requests
typing
websocket-client
asyncio